feat: support dir option

+++ packages/react-email/source/utils/watcher.ts
@@ -1,6 +1,5 @@
-import chokidar from 'chokidar';
+import chokidar, { FSWatcher } from 'chokidar';
 import {
-  CLIENT_EMAILS_PATH,
   CURRENT_PATH,
   EVENT_FILE_DELETED,
   PACKAGE_EMAILS_PATH,
@@ -10,17 +9,21 @@ import fs from 'fs';
 import path from 'path';
 import copy from 'cpy';
 
-export const watcherInstance = chokidar.watch(CLIENT_EMAILS_PATH, {
-  ignoreInitial: true,
-  cwd: CURRENT_PATH,
-  ignored: /(^|[\/\\])\../,
-});
+export const createWatcherInstance = (watchDir: string) =>
+  chokidar.watch(watchDir, {
+    ignoreInitial: true,
+    cwd: CURRENT_PATH,
+    ignored: /(^|[\/\\])\../,
+  });
 
-export const watcher = () =>
+export const watcher = (watcherInstance: FSWatcher, watchDir: string) => {
   watcherInstance.on('all', async (event, filename) => {
-    if (event === EVENT_FILE_DELETED) {
-      const file = filename.split(path.sep);
+    const file = filename.split(path.sep);
+    if (file[1] === undefined) {
+      return;
+    }
 
+    if (event === EVENT_FILE_DELETED) {
       if (file[1] === 'static') {
         if (file[2]) {
           await fs.promises.rm(
@@ -31,17 +34,15 @@ export const watcher = () =>
       }
 
       await fs.promises.rm(path.join(REACT_EMAIL_ROOT, filename));
-    } else {
-      const file = filename.split(path.sep);
-
-      if (file[1] === 'static') {
-        await copy(
-          `${CLIENT_EMAILS_PATH}/static/${file[2]}`,
-          `${REACT_EMAIL_ROOT}/public/static`,
-        );
-        return;
-      }
-
-      await copy(`${CLIENT_EMAILS_PATH}/${file[1]}`, PACKAGE_EMAILS_PATH);
+      return
     }
+    if (file[1] === 'static') {
+      if (file[2]) {
+        const srcPath = path.join(watchDir, 'static', file[2]);
+        await copy(srcPath, `${REACT_EMAIL_ROOT}/public/static`);
+      }
+      return;
+    }
+    await copy(path.join(watchDir, file[1]), PACKAGE_EMAILS_PATH);    
   });
+};
